New Updates

+++ Persistence/Repositories/Product/ProductColerdRepository.cs
@@ -1,4 +1,6 @@
 ﻿using Domain.Entities;
+using Domain.Enums;
+using Domain.Exceptions;
 using Domain.Repositories;
 using Microsoft.EntityFrameworkCore;
 using Persistence.Context;
@@ -12,16 +14,30 @@ namespace Persistence.Repositories
         public ProductColerdRepository(ApplicationDbContext dbContext) =>
            _dbContext = dbContext;
 
-      
+        private bool CanRemove(ColoredProduct product)
+        {
+            var varients = product.Varients.Select(v => v.ProductBelongToOrders);
+            var count = varients.Select(v => v.Count(v =>
+                v.Order.State == OrderStates.Pending
+                || v.Order.State == OrderStates.Confirmed));
+
+            return count.Sum() == 0;
+        }
+
+
         public void Add(ColoredProduct entity) =>
             _dbContext.ColoredProducts.Add(entity);
 
         public void AddRange(List<ColoredProduct> coloredProductVarients) =>
             _dbContext.ColoredProducts.AddRange(coloredProductVarients);
 
-        public void Delete(ColoredProduct entity) =>
-            _dbContext.ColoredProducts.Remove(entity);
-
+        public void Delete(ColoredProduct entity)
+        {
+            if (CanRemove(entity))
+                _dbContext.ColoredProducts.Remove(entity);
+            else
+                throw new NotAllowedException("This Colored Product has Orders in Progress");
+        }
         public void DeleteRangee(List<ColoredProduct> coloredProductVarients) =>
             _dbContext.ColoredProducts.RemoveRange(coloredProductVarients);
